4. Формы. Загрузка файла на сервер. [#504855]

+++ src/main/java/ru/job4j/dreamjob/service/SimpleCandidateService.java
@@ -2,7 +2,9 @@ package ru.job4j.dreamjob.service;
 
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
 import ru.job4j.dreamjob.model.Candidate;
+import ru.job4j.dreamjob.model.Vacancy;
 import ru.job4j.dreamjob.repository.CandidateRepository;
 
 import java.util.Collection;
@@ -14,23 +16,45 @@ public class SimpleCandidateService implements CandidateService {
 
     private final CandidateRepository repository;
 
-    public SimpleCandidateService(CandidateRepository repository) {
+    private final FileService fileService;
+
+    public SimpleCandidateService(CandidateRepository repository, FileService fileService) {
         this.repository = repository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Candidate save(Candidate candidate) {
+    public Candidate save(Candidate candidate, FileDto image) {
+        saveNewFile(candidate, image);
         return repository.save(candidate);
+    }
+
+    private void saveNewFile(Candidate candidate, FileDto image) {
+        var file = fileService.save(image);
+        candidate.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return repository.deleteById(id);
+        var optional = findById(id);
+        if (optional.isPresent()) {
+            repository.deleteById(id);
+            fileService.deleteById(optional.get().getFileId());
+        }
+        return optional.isPresent();
     }
 
     @Override
-    public boolean update(Candidate candidate) {
-        return repository.update(candidate);
+    public boolean update(Candidate candidate, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return repository.update(candidate);
+        }
+        var oldFile = candidate.getFileId();
+        saveNewFile(candidate, image);
+        var isUpdated = repository.update(candidate);
+        fileService.deleteById(oldFile);
+        return isUpdated;
     }
 
     @Override
